requirements.txt
selenium>=4.18.1
webdriver-manager>=4.0.1
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0
openai>=1.0.0
anthropic>=0.18.0 